requirements.txt
# Core Web Framework - Stable and compatible versions
fastapi>=0.100.0,<0.110.0
uvicorn[standard]>=0.23.0,<0.25.0
pydantic>=2.0.0,<3.0.0
python-multipart>=0.0.6
aiofiles>=23.0.0

# Database - Flexible versions
sqlalchemy>=2.0.0,<2.1.0
alembic>=1.12.0,<1.14.0
asyncpg>=0.28.0,<0.30.0

# Async HTTP Client
httpx>=0.24.0,<0.26.0
requests>=2.31.0

# Task Queue & Caching
redis>=4.5.0,<6.0.0
celery>=5.3.0,<5.4.0

# Healthcare & FHIR - Latest stable
fhir.resources>=7.0.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0

# Configuration & Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0
jinja2>=3.1.0
python-dateutil>=2.8.0

# Development & Testing - Flexible
pytest>=7.4.0,<8.0.0
pytest-asyncio>=0.21.0
black>=23.0.0

# Basic AI/ML - Conservative versions to avoid conflicts
openai>=1.0.0,<2.0.0
anthropic>=0.7.0,<1.0.0

# Data Processing
pandas>=2.0.0,<3.0.0
numpy>=1.24.0,<2.0.0

# Optional: Basic NLP without heavy dependencies
nltk>=3.8.0

# Integration APIs - Flexible
docker>=6.0.0,<7.0.0
kubernetes>=28.0.0,<29.0.0

# Image Processing
pillow>=10.0.0
qrcode>=7.4.0
